+++ CHANGELOG.md
@@ -28,6 +28,7 @@ ENHANCEMENT:
 - sequentium.sequence.sequences.integer.recursive.py: added sequence A001608
 - sequentium.sequence.sequences.integer.explicit.py: added sequence A000566 (heptagonal numbers)
 - sequentium.sequence.sequences.integer.explicit.py: added sequence A000567 (octagonal numbers)
+- sequentium.sequence.sequences.integer.explicit.py: added sequence A001106 (nonagonal numbers)
 
 ---
 ## [0.0.1] - 2024-01-02